core: move PhysicsPath to new path lib
