+++ core/envelope-sim/build.gradle
@@ -19,9 +19,11 @@ dependencies {
     // PLEASE ADD AND UPDATE DEPENDENCIES USING libs.versions.toml
 
     implementation project(':osrd-reporting')
+    implementation project(":kt-osrd-path")
     api project(":kt-osrd-sim-infra")
     api project(":osrd-railjson")
     api project(":kt-osrd-utils")
+    testFixturesImplementation project(":kt-osrd-path")
     implementation libs.guava
     implementation libs.slf4j
